rocjitsu: ISA correctness fixes (i24 overflow, transcendental flush, rcp_f16, cvt rounding)
- v_mul_i32_i24 / v_mad_i32_i24: widen to int64 before multiply
- v_rcp_f16: use transcendental::rcp_f32 instead of 1.0f/s
- DS transpose reads: override lane_mask to all-ones per SP3 spec
- Transcendental denormal flushing for rcp, rsq, exp, log (f32)
- v_cvt_nearest/floor_i32_f32: RDNA4 dtype mapping

+++ experimental/rocjitsu/lib/rocjitsu/src/rocjitsu/isa/arch/amdgpu/cdna4/vop3.cpp
@@ -2332,7 +2332,7 @@ void VMadI32I24Vop3::execute_impl(amdgpu::Wavefront &wf) {
     int32_t a = static_cast<int32_t>(src0.read_lane(wf, lane) << 8) >> 8;
     int32_t b = static_cast<int32_t>(src1.read_lane(wf, lane) << 8) >> 8;
     int32_t c = static_cast<int32_t>(src2.read_lane(wf, lane));
-    vdst.write_lane(wf, lane, static_cast<uint32_t>(a * b + c));
+    vdst.write_lane(wf, lane, static_cast<uint32_t>(static_cast<int64_t>(a) * b + c));
   }
 }
